Cleaned up interface, added library stopping
The interface and class arrangement is now cleaner for ClickTrack. In the
future, wrappers for objects will themselves be a nested Java class (see
SubtractiveSynth), which makes the code easier to read as well.

Further, ClickTrack must be explictly stopped and started before processing can
occur. This is because stopping the library is nessecary to release a wakelock